refactor(web): centralize workspace analysis logic in authentication components
- Introduced a new hook, useWorkspaceAnalysis, to streamline workspace analysis across Passcode, Password, and Workspace components.
- Removed redundant memoization in individual components, enhancing code clarity and reducing duplication.
- Improved logic for determining auto-submit conditions based on workspace and team structure.
